% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePPTList_biblio.R
\name{cocicPlot}
\alias{cocicPlot}
\title{Create Co-citation Network}
\usage{
cocicPlot(M, fields = "Authors", n = 50, seed = 1234, ...)
}
\arguments{
\item{M}{is a bibliographic data frame obtained by the converting function convert2df.}

\item{fields}{character One of c("Authors","Papers","Sources")}

\item{n}{The number of items to select.}

\item{seed}{numeric random seed}

\item{...}{Further arguments to be passed to networkPlot}
}
\description{
Create Co-citation Network
}
\examples{
\dontrun{
net=cocicPlot(M,"Authors",n=50,verbose=FALSE)
net=cocicPlot(M,"Papers",n=50,verbose=FALSE)
net=cocicPlot(M,"Sources",n=30,verbose=FALSE)
plot(net$graph)
}
}
