% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePPTList_biblio.R
\name{cocPlot}
\alias{cocPlot}
\title{Create keyword co-occurrences network}
\usage{
cocPlot(M, fields = "ID", n = 50, seed = 1234, ...)
}
\arguments{
\item{M}{is a bibliographic data frame obtained by the converting function convert2df.}

\item{fields}{character  One of c("ID","DE","TI","AB")}

\item{n}{The number of items to select.}

\item{seed}{numeric random seed}

\item{...}{Further arguments to be passed to networkPlot}
}
\description{
Create keyword co-occurrences network
}
\examples{
\dontrun{
net=cocPlot(M,"ID",n=50,verbose=FALSE)
plot(net$graph)
}
}
