% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePPTList_biblio.R
\name{collaboPlot}
\alias{collaboPlot}
\title{Create keyword collaboration network}
\usage{
collaboPlot(M, fields = "Authors", n = 50, seed = 1234, ...)
}
\arguments{
\item{M}{is a bibliographic data frame obtained by the converting function convert2df.}

\item{fields}{character One of c("Authors","Institutions","Countries")}

\item{n}{The number of items to select.}

\item{seed}{numeric random seed}

\item{...}{Further arguments to be passed to networkPlot}
}
\description{
Create keyword collaboration network
}
\examples{
\dontrun{
net=collaboPlot(M,"Authors",n=50,verbose=FALSE)
net=collaboPlot(M,"Institutions",n=50,verbose=FALSE)
net=collaboPlot(M,"Countries",n=30,verbose=FALSE)
plot(net$graph)
}
}
