% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePPTList_biblio.R
\name{countrycollaboration}
\alias{countrycollaboration}
\title{Make a world map plot}
\usage{
countrycollaboration(
  M,
  label = FALSE,
  edgesize = 2.5,
  min.edges = 2,
  log = TRUE,
  guidefill = FALSE
)
}
\arguments{
\item{M}{is a bibliographic data frame obtained by the converting function convert2df.}

\item{label}{logical Whether or not show label}

\item{edgesize}{numeric}

\item{min.edges}{numeric}
}
\description{
Make a world map plot
}
