% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePPTList_biblio.R
\name{degreePlot}
\alias{degreePlot}
\title{draw degree plot}
\usage{
degreePlot(net)
}
\arguments{
\item{net}{A list as a result of networkPlot}
}
\description{
draw degree plot
}
\examples{
\dontrun{
net=cocPlot(M,"ID",n=50)
degreePlot(net)
}
}
