% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePPTList_biblio.R
\name{hindexAuthor}
\alias{hindexAuthor}
\title{get hindex data about author}
\usage{
hindexAuthor(M, results, k = 20)
}
\arguments{
\item{M}{is a bibliographic data frame obtained by the converting function convert2df.}

\item{results}{Result of biblioAnalysis}

\item{k}{integer}
}
\description{
get hindex data about author
}
