% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePPTList_biblio.R
\name{hindexSource}
\alias{hindexSource}
\title{get hindex data of source}
\usage{
hindexSource(M, k = 20)
}
\arguments{
\item{M}{is a bibliographic data frame obtained by the converting function convert2df.}

\item{k}{integer}
}
\description{
get hindex data of source
}
