% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePPTList_biblio.R
\name{locCit}
\alias{locCit}
\title{Local citation score}
\usage{
locCit(M, k = 20)
}
\arguments{
\item{M}{is a data frame obtained by the converting function convert2df and modified with modifyBiblioData}

\item{k}{numeric}
}
\description{
Local citation score
}
