% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePPTList_biblio.R
\name{makePPTList_biblio2}
\alias{makePPTList_biblio2}
\title{Make second Powerpoint List with biblio data}
\usage{
makePPTList_biblio2(
  filepath,
  field1 = "Authors",
  n1 = 50,
  seed1 = 1234,
  field2 = "Authors",
  n2 = 50,
  seed2 = 1234,
  field3 = "ID",
  n3 = 50,
  seed3 = 1234,
  method = "MCA",
  field4 = "ID",
  minDegree = 40
)
}
\arguments{
\item{filepath}{character file path}

\item{field1}{Character to be passed to collaboPlot}

\item{n1}{numeric to be passed to collaboPlot}

\item{seed1}{numeric random seed to be passed to collaboPlot}

\item{field2}{Character to be passed to cocicPlot}

\item{n2}{numeric to be passed to cocicPlot}

\item{seed2}{numeric random seed to be passed to cocicPlot}

\item{field3}{Character to be passed to cocPlot}

\item{n3}{numeric to be passed to cocPlot}

\item{seed3}{numeric random seed to be passed to cocPlot}

\item{method}{character factorial method. One of c("MCA","CA","MDS")}

\item{field4}{character One of c("ID","DE","ID_TM","DE_TM","TI","AB")}

\item{minDegree}{numeric minimum occurrences of terms to analize and plot.}
}
\description{
Make second Powerpoint List with biblio data
}
