% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePPTList_Pubmed.R
\name{makeQuery}
\alias{makeQuery}
\title{Make query with keyword}
\usage{
makeQuery(key = "", author = "", start = 2000, stop = 2021)
}
\arguments{
\item{key}{character keyword}

\item{author}{name of author}

\item{start}{integer start year}

\item{stop}{integer stop year}
}
\description{
Make query with keyword
}
\examples{
makeQuery(key="",author="Moon KW")
makeQuery(key="machine learning",author="")
}
