% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapworld.R
\name{mapworld}
\alias{mapworld}
\title{Make a world map plot}
\usage{
mapworld(M)
}
\arguments{
\item{M}{is a bibliographic data frame obtained by the converting function convert2df.}
}
\description{
Make a world map plot
}
