% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePPTList_Pubmed.R
\name{metaData2df}
\alias{metaData2df}
\title{Convert xml PubMed bibliographic data into a dataframe}
\usage{
metaData2df(D, removeNoAbstract = TRUE, addPMC = FALSE)
}
\arguments{
\item{D}{is a list following the xml PubMed structure, downloaded using the function pmApiRequest.}

\item{removeNoAbstract}{logical whether or not remove data with no abstract}

\item{addPMC}{logical wheter or not add PMCID}
}
\description{
Convert xml PubMed bibliographic data into a dataframe
}
\examples{
query=makeQuery(key="machine learning")
D <-query2metaData(query)
M <- metaData2df(D)
}
