% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePPTList_biblio.R
\name{MLCSources}
\alias{MLCSources}
\title{Most locally cited source}
\usage{
MLCSources(M, k = 20)
}
\arguments{
\item{M}{is a data frame obtained by the converting function convert2df and modified with modifyBiblioData}

\item{k}{numeric}
}
\description{
Most locally cited source
}
