% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePPTList_biblio.R
\name{modifyBiblioData}
\alias{modifyBiblioData}
\title{modify biblioData}
\usage{
modifyBiblioData(M)
}
\arguments{
\item{M}{is a bibliographic data frame obtained by the converting function convert2df.}
}
\description{
modify biblioData
}
