% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPCplot.R
\name{MPCplot}
\alias{MPCplot}
\title{Make mpost productive countries plot}
\usage{
MPCplot(
  sumResult,
  n = 20,
  min = 2,
  legend.position = "bottom",
  title = "Most Productive Countries"
)
}
\arguments{
\item{sumResult}{A list as a result of summary of biblioAnalysis()}

\item{n}{integer Number of conuntries to plot}

\item{min}{minimum number of publication to label}

\item{legend.position}{character specifying legend position}

\item{title}{character title of plot}
}
\value{
A ggplot
}
\description{
Make mpost productive countries plot
}
\examples{
data(scientometrics, package = "bibliometrixData")
results<-biblioAnalysis(scientometrics)
sumResult<-summary(results,k=20)
MPCplot(sumResult,n=10)
MPCplot(sumResult)
MPCplot(sumResult,n=30,min=10)
}
