% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePPTList_biblio.R
\name{myCLEplot}
\alias{myCLEplot}
\title{draw cleveland plot}
\usage{
myCLEplot(data, y = 1, x = 2, xname = NULL, yname = NULL, digits = 2)
}
\arguments{
\item{data}{A data.frame}

\item{y}{numeric position of y-axis variable}

\item{x}{numeric position of x-axis variable}

\item{xname}{character Name of x-axis variable}

\item{yname}{character Name of y-axis variable}

\item{digits}{integer indicating the number of decimal places}
}
\description{
draw cleveland plot
}
