% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myPmApiRequest.R
\name{myPmApiRequest}
\alias{myPmApiRequest}
\title{Gather bibliographic content from PubMed database using NCBI entrez APIs}
\usage{
myPmApiRequest(query, limit, api_key = NULL)
}
\arguments{
\item{query}{is a character. It contains a search query formulated using the Entrez query language.}

\item{limit}{numeric. It indicates the max number of records to download.}

\item{api_key}{character. It contains a valid api key API keys for the NCBI E-utilities.}
}
\description{
It gathers metadata about publications from the NCBI PubMed database. The use of NCBI PubMed APIs is entirely free, and doesn't necessary require an API key. The function pmApiRequest queries NCBI PubMed using an entrez query formulated through the function pmQueryBuild.
}
