% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myReadRDS.R
\name{myReadRDS}
\alias{myReadRDS}
\title{Read RDS file with unique id}
\usage{
myReadRDS(dataname, id = "")
}
\arguments{
\item{dataname}{string dataname}

\item{id}{Name of id field}
}
\description{
Read RDS file with unique id
}
