% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePPTList_biblio.R
\name{myThreeFieldsPlot}
\alias{myThreeFieldsPlot}
\title{Draw threeFiledsPlot with proprocessinf data}
\usage{
myThreeFieldsPlot(
  M,
  fields = c("AU", "DE", "SO"),
  n = c(20, 20, 20),
  width = 1200,
  height = 600
)
}
\arguments{
\item{M}{is a data frame obtained by the converting function convert2df and modified with modifyBiblioData}

\item{fields}{is a character vector. It indicates the fields to analyze using the standard WoS field tags.}

\item{n}{is a integer vector. It indicates how many items to plot, for each of the three fields. Default is n = c(20, 20, 20)}

\item{width}{is an integer. It indicates the plot width (in pixel). Default is width=1200.}

\item{height}{is an integer. It indicates the plot height (in pixel). Default is height=600.}
}
\description{
Draw threeFiledsPlot with proprocessinf data
}
