% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePPTList_biblio.R
\name{net2Data}
\alias{net2Data}
\title{Extract data from co-occurence network plot}
\usage{
net2Data(net)
}
\arguments{
\item{net}{result of networkPlot}
}
\description{
Extract data from co-occurence network plot
}
\examples{
\dontrun{
net=cocPlot(M,"ID",n=50,verbose=FALSE)
net2Data(net)
}
}
