% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePPTList_Pubmed.R
\name{plotWC}
\alias{plotWC}
\title{Make Wordcloud}
\usage{
plotWC(df, type = 1, seed = 1234, ...)
}
\arguments{
\item{df}{A data.frame}

\item{type}{integer}

\item{seed}{integer}

\item{...}{further arguments}
}
\description{
Make Wordcloud
}
\examples{
query=makeQuery(key="machine learning")
D<-query2metaData(query)
M<-metaData2df(D)
df=df2clean(M)
plotWC(df)
}
