% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordlist.R
\name{plotWC1}
\alias{plotWC1}
\title{Make word cloud}
\usage{
plotWC1(M, type = 1)
}
\arguments{
\item{M}{is a bibliographic data frame obtained by the converting function convert2df.}

\item{type}{integer}
}
\description{
Make word cloud
}
