% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMCids.R
\docType{data}
\name{PMCids}
\alias{PMCids}
\title{Pubmed Central ids data from NIH}
\format{
The data are 7,029,270 observations and 3 variables:
\describe{
  \item{PMID}{use simple numbers, e.g., 23193287}
  \item{PMCID}{include the ‘PMC’ prefix, e.g., PMC3531190}
  \item{DOI}{DOI}
}
}
\usage{
PMCids
}
\description{
Pubmed Central ids data from NIH
}
\keyword{datasets}
