% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmid2pmc.R
\name{pmid2pmc}
\alias{pmid2pmc}
\title{Get pmc id with pmids}
\usage{
pmid2pmc(pmids)
}
\arguments{
\item{pmids}{String A character vector of pmids}
}
\description{
Get pmc id with pmids
}
\examples{
pmids=c("10848627","111111111","10848628")
df=pmid2pmc(pmids)
}
