% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmedModule.R
\name{pubmedModule}
\alias{pubmedModule}
\title{Server function of sTree shiny module}
\usage{
pubmedModule(input, output, session, PPTdata)
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{session}{session}

\item{PPTdata}{PPTdata}
}
\description{
Server function of sTree shiny module
}
