% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePPTList_Pubmed.R
\name{query2clean}
\alias{query2clean}
\title{Make clean data from query}
\usage{
query2clean(query, limit = 200)
}
\arguments{
\item{query}{character}

\item{limit}{integer}
}
\description{
Make clean data from query
}
\examples{
query=makeQuery(key="machine learning")
query2clean(query)
}
