% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePPTList_Pubmed.R
\name{query2metaData}
\alias{query2metaData}
\title{retrieve pubmed data from query}
\usage{
query2metaData(query, limit = 200)
}
\arguments{
\item{query}{character}

\item{limit}{integer}
}
\description{
retrieve pubmed data from query
}
\examples{
query=makeQuery(key="propensity score")
D <-query2metaData(query)
}
