% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePPTList_biblio.R
\name{removeGeoms}
\alias{removeGeoms}
\title{Remove selected layers of ggplot}
\usage{
removeGeoms(p, choices = NULL, geoms = NULL, last = TRUE)
}
\arguments{
\item{p}{A ggplot}

\item{choices}{numeric}

\item{geoms}{character}

\item{last}{logical}
}
\description{
Remove selected layers of ggplot
}
