% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riverplotTE.R
\name{riverplotTE}
\alias{riverplotTE}
\title{Make riverplot for thematic evolution}
\usage{
riverplotTE(
  Nodes,
  Edges,
  measure = "inclusion",
  nodewidth = 10,
  textcex = 1,
  ...
)
}
\arguments{
\item{Nodes}{is a list of nodes obtained by thematicEvolution function.}

\item{Edges}{is a list of edges obtained by thematicEvolution function.}

\item{measure}{is a character. It can be measure=("inclusion","stability", "weighted")}

\item{nodewidth}{width of the node (relative to font size)}

\item{textcex}{size of text}

\item{...}{further arguments to be passed to riverplot}
}
\description{
Make riverplot for thematic evolution
}
\examples{
library(bibliometrix)
library(riverplot)
data(scientometrics, package = "bibliometrixData")
years=c(2000,2012)
nexus <- thematicEvolution(scientometrics,field="ID",years=years,n=100,minFreq=2)
par(mfrow=c(1,1))
riverplotTE(nexus$Nodes,nexus$Edges)
}
