% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePPTList_biblio.R
\name{sourceGrowthPlot}
\alias{sourceGrowthPlot}
\title{source growth Plot}
\usage{
sourceGrowthPlot(...)
}
\arguments{
\item{...}{arguments to be passed to sourceGrowth}
}
\description{
source growth Plot
}
\examples{
#sourceGrowthPlot(M,top=5,cdf=FALSE)
}
