% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threeFieldsRiverPlot.R
\name{threeFieldsRiverPlot}
\alias{threeFieldsRiverPlot}
\title{Draw three fields river plot}
\usage{
threeFieldsRiverPlot(res, nodewidth = 3, textcex = 0.7, ...)
}
\arguments{
\item{res}{result of bibliometrix::threeFieldsPlot}

\item{nodewidth}{numeric The width of node}

\item{textcex}{numeric size of node label}

\item{...}{further arguments to be passed to riverplot}
}
\description{
Draw three fields river plot
}
\examples{
library(bibliometrix)
library(riverplot)
data(scientometrics, package = "bibliometrixData")
res=threeFieldsPlot(scientometrics, fields=c("DE","AU","CR"),n=c(10,10,10))
threeFieldsRiverPlot(res,nodewidth=7)
}
