% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePPTList_biblio.R
\name{wordCount}
\alias{wordCount}
\title{Keywords Count}
\usage{
wordCount(M, terms = "ID", k = 20)
}
\arguments{
\item{M}{is a data frame obtained by the converting function convert2df and modified with modifyBiblioData}

\item{terms}{One of c("ID","DE","TI","AB")}

\item{k}{numeric}
}
\description{
Keywords Count
}
