% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePPTList_biblio.R
\name{wordDynamicPlot}
\alias{wordDynamicPlot}
\title{Make wordDynamic Plot}
\usage{
wordDynamicPlot(M, terms = "ID", top = 10, cdf = TRUE)
}
\arguments{
\item{M}{is a data frame obtained by the converting function convert2df and modified with modifyBiblioData}

\item{terms}{One of c("ID","DE","TI","AB")}

\item{top}{numeric}

\item{cdf}{logical Whrther or not cumulative}
}
\description{
Make wordDynamic Plot
}
