% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordlist.R
\name{wordlist}
\alias{wordlist}
\title{Get word list}
\usage{
wordlist(M, Field = "ID", n = 50, measure = "freq", ngrams = 1)
}
\arguments{
\item{M}{is a bibliographic data frame obtained by the converting function convert2df.}

\item{Field}{character}

\item{n}{numeric}

\item{measure}{character}

\item{ngrams}{numeric}
}
\description{
Get word list
}
