% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePPTList_biblio.R
\name{makePPTList_biblio}
\alias{makePPTList_biblio}
\title{Make Powerpoint List with biblio data}
\usage{
makePPTList_biblio(
  filepath,
  all = FALSE,
  main = FALSE,
  sources = FALSE,
  authors = FALSE,
  affiliations = FALSE,
  country = FALSE,
  documents = FALSE,
  keywords = FALSE
)
}
\arguments{
\item{filepath}{character file path}

\item{all}{logical}

\item{main}{logical}

\item{sources}{logical}

\item{authors}{logical}

\item{affiliations}{logical}

\item{country}{logical}

\item{documents}{logical}

\item{keywords}{logical}
}
\description{
Make Powerpoint List with biblio data
}
