% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePPTList_Pubmed.R
\name{makePPTList_pubmed}
\alias{makePPTList_pubmed}
\title{Make Powerpoint list for pubmed wordcloud}
\usage{
makePPTList_pubmed(query = "", seed = 1234)
}
\arguments{
\item{query}{character query}

\item{seed}{numeric seed number}
}
\description{
Make Powerpoint list for pubmed wordcloud
}
\examples{
query=makeQuery(key="machine learning")
result=makePPTList_pubmed(query=query)
}
