% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePPTList_Pubmed.R
\name{query2clean}
\alias{query2clean}
\title{retrieve pubmed data from query and make clean data}
\usage{
query2clean(query)
}
\arguments{
\item{query}{character}
}
\description{
retrieve pubmed data from query and make clean data
}
\examples{
\dontrun{
query=makeQuery(key="propensity score")
query2clean(query)
}
}
