% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePPTList_biblio.R
\name{wordCount}
\alias{wordCount}
\title{Keywords Count}
\usage{
wordCount(M, k = 100)
}
\arguments{
\item{M}{is a data frame obtained by the converting function convert2df and modified with modifyBiblioData}

\item{k}{numeric}
}
\description{
Keywords Count
}
